functions/health_checker/requirements.txt
functions-framework==3.8.2
google-cloud-bigquery==3.27.0
google-cloud-firestore==2.19.0
google-cloud-logging==3.14.0
firebase-admin==6.5.0